feat: add ProjectFormDialog for project creation and editing functionality
